+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Remote Procedure Calls for .NET Changelog
 
+## <a name="3.3.8-3.3.10"></a> 3.3.8-3.3.10 (2020-12-25)
+### Features
+* Extended RestClient with PATCH method (only for netstandard2.1)
+* Added route builder helper
+
 ## <a name="3.3.7"></a> 3.3.7 (2020-12-11)
 
 ### Features